+++ src/Components/Schema/index.tsx
@@ -1,63 +1,29 @@
-import { useState, useEffect, useCallback, memo } from 'react';
+import { useState } from 'react';
+import { Button, Drawer, Stack, Divider } from '@mui/material';
+import SchemaList from './SchemaList';
 
-import {
-  Box,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemText,
-  Typography,
-  Button,
-} from '@mui/material';
-import { useQuery } from '@apollo/client';
-import { SCHEMA } from '../../apollo/schema';
-
-type SchemaType = {
-  name: string;
-  description: string;
-  type: { name: string };
-};
-
-const Schema: React.FC<{ name: string }> = ({ name }) => {
+const Schema = () => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
-
-  const { data } = useQuery(SCHEMA, {
-    variables: { name },
-  });
-  console.log(data);
 
   return (
     <>
       <Button variant='contained' onClick={() => setIsOpen(!isOpen)}>
         Docs
       </Button>
-      {isOpen && (
-        <Box>
-          <Typography variant='h5'>{data.__type.name}</Typography>
-          <List>
-            {data.__type.fields.map(
-              ({ name, description, type }: SchemaType) => (
-                <ListItem key={name}>
-                  <ListItemButton>
-                    <ListItemText
-                      primary={
-                        <Typography variant='h5'>
-                          {`${name}: ${type.name ?? ''}`}
-                        </Typography>
-                      }
-                      secondary={
-                        <Typography variant='subtitle1'>
-                          {description}
-                        </Typography>
-                      }
-                    ></ListItemText>
-                  </ListItemButton>
-                </ListItem>
-              )
-            )}
-          </List>
-        </Box>
-      )}
+      <Drawer
+        anchor='right'
+        open={isOpen}
+        onClose={() => setIsOpen(!isOpen)}
+        variant='persistent'
+      >
+        <Stack
+          direction='row'
+          spacing={2}
+          divider={<Divider orientation='vertical' />}
+        >
+          <SchemaList name='Query' />
+        </Stack>
+      </Drawer>
     </>
   );
 };
